+++ src/utils.js
@@ -1,0 +1,30 @@
+const fs = require("fs");
+
+const isObject = function (o) {
+  return o === Object(o) && !Array.isArray(o) && typeof o !== "function";
+};
+
+const isString = (s) => typeof s === "string";
+
+function saveJson(json, filename) {
+  try {
+    const data = JSON.stringify(json);
+    fs.writeFileSync(filename, data);
+  } catch (e) {
+    console.error("Failed to save JSON:", filename, e);
+  }
+}
+
+function loadJson(filename, ignore) {
+  try {
+    let rawData = fs.readFileSync(filename);
+    return JSON.parse(rawData);
+  } catch (e) {
+    if (!ignore) {
+      console.error("Failed to load JSON:", filename, e);
+    }
+  }
+  return null;
+}
+
+module.exports = { saveJson, loadJson, isObject, isString };
